Debugging STM32F7 version.

+++ lathe_src/i2cx.cpp
@@ -66,6 +66,7 @@ unsigned int millis();
 #if defined(STM32F1) || defined(STM32F4)
 void i2cWrite(uint8_t);
 #endif	/* STM32F1 || STM32F4 */
+
 void i2cWrite(uint8_t *data, uint32_t size);
 
 void i2cWaitBusy();
@@ -238,7 +239,7 @@ void i2cWrite(uint8_t data)
 }
 #endif
 
-void i2cWrite(uint8_t *data, uint16_t size)
+void i2cWrite(uint8_t *data, uint32_t size)
 {
  unsigned int timeout = 20;
  unsigned int start = millis();
